chatbot_tutorial/requirements.txt
langchain[openai]
python-dotenv
langgraph
IPython
